+++ ControleAcesso/src/br/ufsc/ine5605/controleacesso/Model/Aluno.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
  */
 public class Aluno extends Pessoa{
     private String curso;
-    public Aluno(int matricula, String nome, int telefone, String email, String curso) {
+    public Aluno(int matricula, String nome, long telefone, String email, String curso) {
         super(matricula, nome, telefone, email);
         setCurso(curso);
     }
