+++ ControleAcesso/src/br/ufsc/ine5605/controleacesso/interfaces/ICtrlPessoa.java
@@ -24,9 +24,9 @@ public interface ICtrlPessoa {
      * @param telefone - Qual telefone do aluno.
      * @param nome - Nome do aluno a ser cadastrado.
      * @param email - email do aluno.
-     
+     *
      */
-    public boolean incluiAluno(int matricula, String nome, int telefone, String email, String curso);
+    public boolean incluiAluno(int matricula, String nome, long telefone, String email, String curso);
 
     /**
      * É utilizado pela TelaPessoa para incluir um novo Servidor na lista de
@@ -42,19 +42,20 @@ public interface ICtrlPessoa {
      * @param listaSalas - Lista de salas (salas cadastradas na tela de salas) o
      * aluno tem acesso,.
      */
-    public boolean incluiServidor(int matricula, String nome, int telefone, String email, String cargo, boolean administrador);
+    public boolean incluiServidor(int matricula, String nome, long telefone, String email, String cargo, boolean administrador);
 
     /**
      * É utilizado pela TelaPessoa para deletar uma pessoa na lista de pessoas.
      *
      * @param matricula - Qual matricula o aluno possui.
-     * @return 
+     * @return
      */
     public boolean delPessoa(int matricula);
-    
-    public boolean alteradorDeCadastroAluno(int matricula, String nome, int telefone, String email, String curso);
-    public boolean alteradorDeCadastroServidor(int matricula, String nome, int telefone, String email, String cargo, boolean administrador);
-    
+
+    public boolean alteradorDeCadastroAluno(int matricula, String nome, long telefone, String email, String curso);
+
+    public boolean alteradorDeCadastroServidor(int matricula, String nome, long telefone, String email, String cargo, boolean administrador);
+
     /**
      * É utilizado pela TelaPessoa para incluir uma sala cadastrada na lista de
      * salas da pessoa.
@@ -62,7 +63,7 @@ public interface ICtrlPessoa {
      * @param matricula
      * @param codigoSala - código na listaSalas da pessoa a sala encontrada com
      * esse codigo sala.
-     * @return 
+     * @return
      */
     public boolean cadastraSalaNaPessoa(int matricula, String codigoSala);
 
@@ -71,7 +72,7 @@ public interface ICtrlPessoa {
      * @param matricula
      * @param codigoSala - é utilizado para encontrara a sala na lista de salas
      * do aluno e deleta a sala da lista.
-     * @return 
+     * @return
      */
     public boolean delSalaNaPessoa(int matricula, String codigoSala);
 
@@ -81,12 +82,14 @@ public interface ICtrlPessoa {
      * @return - deve retonar um arrayList com as salas que a pessoa possui em
      * seu cadastro.
      */
-    public String listaSalasCadastradas(int matricula) ;
+    public String listaSalasCadastradas(int matricula);
+
     /**
      * Metodo responsável por mostrar todas as pessoas cadastradas no sistema
+     *
      * @return - retorna uma lista de pessoas em String
      */
     public String listAllPessoasCadastradas();
-    
+
     public Pessoa findPessoaByMatricula(int matricula);
 }
